Fix external buffers cursor test
The test was previously failing due to the assumption that the
local-options help text will always be around line 190 of the help file,
which seems to not be true anymore. Avoid that assumption by checking
that the cursors for VS Code and Neovim are in sync and that they're
both pointing to the correct line.